(BSR)[API] test: Move all User-related tests to existing module in tests/core/users/
